od: Changed od_int64 to int.
Summary: As int is sufficient for practical purposes.

Test Plan: Compiles.

Reviewers: nanne

Reviewed By: nanne

Subscribers: helene, raman

Maniphest Tasks: T237

Differential Revision: https://phabricator.dgbes.com:8443/D7337

git-svn-id: https://svn.opendtect.org/od/trunk@38551 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/EarthModel/emfault3dpainter.cc
@@ -277,8 +277,7 @@ bool Fault3DPainter::paintStickOnRLine( const Geometry::FaultStickSurface& fss,
 	bid = SI().transform( pos.coord() );
 	const TrcKey trckey = Survey::GM().traceKey(
 		Survey::GM().default3DSurvID(), bid.inl(), bid.crl() );
-	const int idx = (int)path_->indexOf( trckey );
-	    //TODO the cast above violates the design
+	const int idx = path_->indexOf( trckey );
 	if ( idx < 0 ) continue;
 
 	Coord3 editnormal( getNormalInRandLine(idx), 0 );
@@ -453,8 +452,7 @@ void Fault3DPainter::genIntersectionAuxData( EM::Fault3D& f3d,
 	{
 	    const TrcKey trckey = Survey::GM().traceKey(
 		    Survey::GM().default3DSurvID(),posbid.inl(),posbid.crl() );
-	    const int trcidx = (int)path_->indexOf( trckey );
-		//TODO the cast above violates the design
+	    const int trcidx = path_->indexOf( trckey );
 	    if ( trcidx != -1 )
 	    {
 		const double z = zat ? zat->transform(pos) : pos.z;
