+++ src/visSurvey/visseis2ddisplay.cc
@@ -419,8 +419,7 @@ void Seis2DDisplay::updateChannels( int attrib, TaskRunner* taskr )
 		    const int trcnr =
 			trcdisplayinfo_.alltrcnrs_[crlidx+startidx];
 		    const TrcKey trckey = Survey::GM().traceKey(geomid_,trcnr);
-		    const int trcidx = (int)seisdp->getGlobalIdx( trckey );
-			//TODO the cast above violates the design
+		    const int trcidx = seisdp->getGlobalIdx( trckey );
 		    const float* trcptr = seisdp->getTrcData( idx, trcidx );
 		    OffsetValueSeries<float> trcstor =
 				seisdp->getTrcStorage( idx, trcidx );
@@ -932,8 +931,7 @@ bool Seis2DDisplay::getCacheValue( int attrib, int version,
 
     const int trcnr = geometry_.positions()[traceidx].nr_;
     const TrcKey trckey = Survey::GM().traceKey( geomid_, trcnr );
-    const int trcidx = (int)regsdp->getGlobalIdx( trckey );
-	    //TODO the cast above violates the design
+    const int trcidx = regsdp->getGlobalIdx( trckey );
     const int sampidx =  regsdp->getZRange().nearestIndex( pos.z );
     const Array3DImpl<float>& array = regsdp->data( version );
     if ( !array.info().validPos(0,trcidx,sampidx) )
